Byte数据类型重解。

+++ src/Main.java
@@ -1,11 +1,14 @@
 import Map.MapTest;
+import dataType.ByteType;
 import dataType.IntegerType;
+import math.BitMath;
 import object.ObjectTest;
 /**
  * Description 主函数入口
  */
 public class Main {
     public static void main(String[] args) {
-        IntegerType.test();
+        ByteType.test();
+
     }
 }
